Change the way the clipping is calculated in curtain mode
When using a navigator, the view in the navigator resize to find
the new minimum rectangle that can encompass all images. When we pass
the new image size without subtracting the clipped part, we get a new
size with the original size plus the clipped part. This causes that
the minimum rectangle found is the maximum size plus the clipped part.

In other words, it does not take into account the new width.